+++ calculator/src/main/java/edu/nextstep/camp/calculator/ui/MemoryAdapter.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator
+package edu.nextstep.camp.calculator.ui
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import edu.nextstep.camp.calculator.R
 import edu.nextstep.camp.calculator.databinding.ItemResultBinding
 import edu.nextstep.camp.data.Memory
